chore: relocate/refactor app entry component
